Move feed class to feeds.py and rename the classes.
